Migrations

+++ src/routes/api/orgs.api.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const appService = require('../../services/app.service');
-const { Org } = require('../../models');
+const { Org, Event } = require('../../models');
 
 router.get('/', async (req, res) => {
     res.send(await Org.find());
@@ -13,7 +13,6 @@ router.get('/:id', async (req, res) => {
 
 router.post('/', async (req, res) => {
     req.body.code = appService.encode(req.body.name);
-    req.body.cityName = req.body.address.city;
     req.body.cityCode = appService.encode(req.body.address.city);
     res.send(await Org(req.body).save());
 });
@@ -21,11 +20,35 @@ router.post('/', async (req, res) => {
 router.put('/:id', async (req, res) => {
     req.body.code = appService.encode(req.body.name);
     req.body.cityCode = appService.encode(req.body.address.city);
-    res.send(await Org.findOneAndUpdate({_id: req.body._id}, req.body));
+
+    let oldOrgCode = appService.encode(req.body.oldName);
+    let oldCityCode = appService.encode(req.body.oldCityName);
+    let org = await Org.findOneAndUpdate({_id: req.body._id}, req.body);
+
+    // Propagate update on events
+    let query = {orgCode: oldOrgCode, cityCode: oldCityCode};
+    let events = await Event.find(query);
+    events.forEach(el => {
+        el.orgCode = req.body.code;
+        el.orgName = req.body.name;
+        el.cityCode = appService.encode(req.body.newCityName);
+        el.cityName = req.body.newCityName;
+        el.save();
+    });
+
+    res.send(org);
 });
 
 router.delete('/:id', async (req, res) => {
-    res.send(await Org.findOneAndDelete({ _id: req.params.id }));
+    // only if there is no event on this org
+    let org = await Org.findById(req.params.id);
+    let event = await Event.findOne({orgCode: org.code});
+    
+    if(!event) {
+        res.send(await Org.findOneAndDelete({ _id: req.params.id }));
+    } else {
+        res.status(400).send({message: 'Remoção de organização não é permitida quando existem eventos cadastrados para ela.'});
+    }
 });
 
 router.get('/check-code/:code/:cityCode', async (req, res) => {
